requirements.txt
django
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
django-encrypted-model-fields
django-cors-headers
